frontend/requirements.txt
streamlit==1.28.0
requests==2.31.0
pillow==10.0.0
numpy==1.24.3
python-multipart==0.0.6